movieData is added
