+++ index.html
@@ -7,6 +7,8 @@
     <link rel="preconnect" href="https://fonts.googleapis.com">
 <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
 <link href="https://fonts.googleapis.com/css2?family=Space+Grotesk:wght@300..700&display=swap" rel="stylesheet">
+<script type="module" src="./js/master.js" defer></script>
+<!-- <script type="module" src="./js/movie-cards/" defer></script> -->
     <title>Movie page</title>
 </head>
 <body>
@@ -21,6 +23,8 @@
       </nav>
     </header>
     <main>
+
+
         <section class="hero">
                 <div class="hero-slideshow">
                     <div class="mover"></div>
@@ -79,6 +83,11 @@
             
             </div>
         </section>
+<section>
+  <div id="movie-cards"></div>
+</section>
+
+
         <section>
           <form action="#" method="post" onsubmit=""> 
             <label for="contact-first-name">
