+++ js/master.js
@@ -1,0 +1,4 @@
+import  appendCards  from "./movie-cards/movie-cards.js";
+ 
+
+ appendCards();
