+++ js/movie-cards/moviedata.js
@@ -1,0 +1,34 @@
+
+
+  const cards = [{
+    id: 1,
+    title:'Echoes of Tomorrow',
+    year:2023,
+    director: 'Manrika',
+    poster: "../../assets/cards/echoes.jpg",
+    synopsis: " In a future where memories can be traded, a rogue scientist uncovers a conspiracy that could rewrite humanity's past and future."
+}, 
+{   id:2,
+    title:'Crimson Horizon' ,
+    year: 2020,
+    director: 'Gaj',
+    poster: "../../assets/cards/crimson.jpg",
+    synopsis: ' small coastal town becomes the epicenter of a mysterious phenomenon as the sea turns blood red, and dark secrets begin to surface.'
+},
+{   id:3,
+    title: 'Shattered Reflections',
+    year:2019,
+    director: 'Jack',
+    poster: "../../assets/cards/reflections.jpg",
+    synopsis: 'A detective haunted by her past must confront her own fractured psyche to solve a murder that hits too close to home.'},
+
+{   id:4,    
+    title: 'The Midnight Paradox',
+        year:2024,
+        director: 'William',
+        poster: "../../assets/cards/paradox.jpg",
+        synopsis:  'When a physicist stumbles upon a way to travel between parallel universes, he faces a race against time to save his world from collapsing.'}]
+
+    
+
+  export default cards
